“kmem address” not working properly when redzone is enabled
When "slub_debug" is enabled with redzoning, "kmem address" does not
work properly. The "red_left_pad" member within "struct kmem_cache" is
currently an "unsigned int", it used to be an "int", but it never was a
"long", hence "red_left_pad" in do_slab_slub() was not initialized
properly. This "red_left_pad" issue resulted in reporting free objects
as "[ALLOCATED]", and in reporting bogus object address when using "set
redzone off".

Signed-off-by: Georges Aureau <georges.aureau@hpe.com>